requirements.txt
ag2==0.9.1.post0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
blinker==1.9.0
capstone==5.0.6
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lief==0.16.6
MarkupSafe==3.0.2
openai==1.82.1
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
